+++ terraform/environments/production/main.tf
@@ -36,5 +36,66 @@ module "s3" {
   source = "../../modules/s3"
 
   bucket_name = var.assets_bucket_name
-  tags        = local.common_tags
+
+  tags = local.common_tags
+}
+
+module "alb" {
+  source = "../../modules/alb"
+
+  project_name = var.project_name
+  environment  = var.environment
+
+  vpc_id            = module.networking.vpc_id
+  public_subnet_ids = module.networking.public_subnet_ids
+  security_group_id = module.security_groups.alb_security_group_id
+
+  tags = local.common_tags
+}
+
+module "cloudfront" {
+  source = "../../modules/cloudfront"
+
+  project_name = var.project_name
+
+  alb_dns_name = module.alb.alb_dns_name
+
+  assets_bucket_regional_domain_name = module.s3.bucket_regional_domain_name
+
+  tags = local.common_tags
+}
+
+resource "aws_s3_bucket_policy" "cloudfront" {
+  depends_on = [module.cloudfront]
+
+  bucket = module.s3.bucket_id
+
+  policy = jsonencode({
+    Version = "2012-10-17"
+
+    Statement = [
+      {
+        Sid    = "AllowCloudFrontServicePrincipal"
+        Effect = "Allow"
+
+        Principal = {
+          Service = "cloudfront.amazonaws.com"
+        }
+
+        Action = [
+          "s3:GetObject"
+        ]
+
+        Resource = [
+          "${module.s3.bucket_arn}/*"
+        ]
+
+        Condition = {
+          StringEquals = {
+            "AWS:SourceArn" = module.cloudfront.distribution_arn
+          }
+        }
+      }
+    ]
+  })
 }