+++ terraform/modules/s3/outputs.tf
@@ -4,8 +4,20 @@ output "bucket_name" {
   value = aws_s3_bucket.assets.bucket
 }
 
+output "bucket_id" {
+  description = "ID of the S3 bucket."
+
+  value = aws_s3_bucket.assets.id
+}
+
 output "bucket_arn" {
   description = "S3 bucket ARN."
 
   value = aws_s3_bucket.assets.arn
+}
+
+output "bucket_regional_domain_name" {
+  description = "Regional domain name of the S3 bucket."
+
+  value = aws_s3_bucket.assets.bucket_regional_domain_name
 }